requirements.txt
pandas
langchain
python-dotenv
openai